requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
openpyxl==3.1.2
scikit-learn==1.2.2
pmdarima==2.0.4
statsmodels==0.14.1
tensorflow==2.13.0
keras-tuner==1.4.6
joblib==1.4.2
